requirements.txt
pandas
numpy
scikit-learn
scikit-surprise
mlflow
dvc
pytest
fastapi
uvicorn
prometheus-client
